ConnectivityManager: callback map is an instance variable
All NetworkCallbacks registered by a user of ConnectivityManager are
stored in a static map that is shared by all instances of
ConnectivityManager for a given process.

This patch changes this map to instead be an instance variable. This
improves isolation during testing by removing the global static state.

Similarly this patch changes the default callback handler to be an
instance variable. The callback handler is instantiated as soon as
a ConnectivityManager instance is created. Note that there is still a
single backing Thread for all ConnectivityManager instances (global
singleton in ConnectivityThread).

Bug: 36701874, 37107940
Test: runtest frameworks-net
Change-Id: Idb0d45ae9c23affd9f30bccbeaa3581c1e803fe2