ConnectivityManager: callback map is an instance variable
All NetworkCallbacks registered by a user of ConnectivityManager are
stored in a static map that is shared by all instances of
ConnectivityManager for a given process.

This patch changes all mutations and reads to this map to go through an
instance variable in ConnectivityManager. This is safe to do because all
operations still lock the same object.

The static singleton map is injected by default as the instance variable
used by all ConnectivityManager instances. This allows to preserve the
existing global state semantics while improving isolation during testing
where independent map instances can be injected.

Similarly this patch changes the default callback handler to be an
instance variable. The callback handler is instantiated as soon as
a ConnectivityManager instance is created. Note that there is still a
single backing Looper for all ConnectivityManager instances (global
singleton in ConnectivityThread).

Bug: 36701874, 37107940
Test: runtest frameworks-net
Change-Id: Idb0d45ae9c23affd9f30bccbeaa3581c1e803fe2