Update patch set 1
Patch Set 1: Code-Review-1

Looks like this could cause a change in behaviour in the case where an app registers a callback on one instance of CM and unregisters it on another. While this is admittedly a corner case (and may even be impossible given the current state of the code because of the use of StaticApplicationContextServiceFetcher), we've had problems with this sort of thing in the past, and I'm not eager to make it easier for us to have those problems again.

Can we do something else that maintains the static in production code but still allows testing? For example, can we provide a separate constructor for use by the tests that takes in a callback map?

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1