Update test API for android.legacy.test
The android.legacy.test library will be added to the SDK for use
by developers after most of the android.test classes are removed
from the public API. As it will be statically included in
developer's applications it should be built against the public
API. Unfortunately, there are a few classes and methods that it
needs which are not public. This adds them.

Bug: 30188076
Test: make checkbuild
Change-Id: I1e8bfd597ff0289b1d44f7f25796997202c4e1be
Merged-In: Id823489519834e0e4da9cd4b18885679abb8d0c5