Update patch set 19
Patch Set 19: Presubmit-Ready+1

Most of the time a network will not get promoted to the default until it validates, so that's fine.

Not requiring validation is important for the occasions where the system promotes a non-validated network to be the default.  And we really should check the time even in those circumstances.

I'm rerequesting a treehugger run.

Patch-set: 19
Reviewer: Gerrit User 1000451 <1000451@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1