Captive portal: take into account slow dns queries
This patch changes the portal probing logic to handle more robustly the
case of slow DNS queries. More specifically the fallback probe 3s wait
is not started before both DNS queries complete. This can otherwise
cause the result of the http probe to be ignored, which may cause the
system to miss captive portals advertising themselves on the url used by
the http probe but not on the url used by the fallback probe.

Bug: 38259299
Bug: 36532213
Test: manually tested with nearby captive portal networks
Change-Id: Iedf8f72b02a7d59b0cf8e4cfabfd595591351d5f