Captive portal: better async handling of fallback probe
This patch improves the asynchronous handling of the http, https and
fallback probes, at the cost of using a third thread for the fallback
probe. That thread is only created and started if neither the http nor
the https probe return a conclusive answer within 3 seconds, which
mitigates the cost of this additional thread.

Using a third thread allows to simplify the logic starting the probes,
waiting for their results, and handling the results. It also makes
the handling of results more robust to probes racing against each other.

Bug: 38259299
Bug: 36532213
Test: manually tested with nearby captive portal networks
Change-Id: Ic446074fa77f84f5bdb27bfccdf144295555b844