Update patch set 1
Patch Set 1:

just a comment; this is just a temporary solution due to the fact that sdcards are not visible due to the fact that fbe does not support adoptable cards, and if adoptable is not set card is not visible.

A better solution would be do disconnect visibility from adoptability, adding visibility configuration to fstab, then it is possible to choose to have cards visible and not support adoptable cards.

Patch-set: 1
CC: Gerrit User 1004424 <1004424@85c56323-6fa9-3386-8a01-6480fb634889>