Reduce app size by downgrading inactive apps
Active apps here refer to the apps which were either active
in foregrond or in background and also used by other packages.
It includes two steps:
a) Apps which are inactive for X days:
   These apps will be downgraded to quicken.
   X is determined by sysprop pm.dexopt.unopt_after_inactive_days
b) Apps which are inactive for Y days:
   odex and vdex files associated with the apps will be removed. When the
   user reuses the app, it will run directly from the dex file.
   Y is determined by sysprop pm.dexopt.fallback_after_inactive_days.

If the system properties are not set, no effect will take place.

The above operations will take place in background dexopt service.
If user uses the app again, it will again be speed-compiled when
background dexopt service starts next time.

Bug: 36598475
Test: adb shell cmd package bg-dexopt-job
Change-Id: I68f9f617d6722a7ba8b00aa2181cb38a165cfc51