Quicken inactive apps in background dexopt service
The apps which will not be used for a threshold period of time
in foreground or in running in background with other apps using it,
will be unoptimized to quicken.

The threshold will be set using system property "pm.dexopt.deopt_threshold".
If the property is not set, it will not quicken the apps.

Bug: 36598475
Test: adb shell cmd package bg-dexopt-job
Change-Id: I68f9f617d6722a7ba8b00aa2181cb38a165cfc51