OMS: add support for user-installed overlays
Teach the package manager to look for overlay packages everywhere, not
just in /vendor/overlay. Implement the necessary callbacks in the
overlay manager (onOverlayPackage.*).

Also, update the overlay manager unit tests to install its own overlays
as part of the setup rather than relying on the user to have done so
before.

Note: this commit allows any overlay package to be used which is a
security risk. A future commit will harden security by restricting which
overlay packages are allowed to take effect.

Test: bit FrameworksServicesTests:com.android.server.om.{OverlayManagerSettingsTests,OverlayManagerTests,RuntimeResourceOverlayTests}
Change-Id: Ib4556b0f65cc27772bc2efb6063ff1963811de3d