Update patch set 16
Patch Set 16:

I have several comments on the formatting of the dumpsys.

It seems like gerrit changed the whitespace when you pasted the dumpsys into a comment. Could you send me an unmodified dump?

In mUserQuotaTracker, each line starts with "Object uid" which seems overly generic. How about we just print it like this:
UID 123: {...}

In mUserQuotaTracker, the values between brackets are a flat, space separated list. This is a bit confusing since you're flattening three objects and putting the object names as tokens in the list. Maybe it should look something like this:
UID 123: {socket{...}, transform{...}, spi{...}}

In mTransformRecords, you wrap a lot of individual values in parentheses. I don't think it adds any clarity to do so.

In mTransformRecords, it may be because gerrit broke your whitespace, but there should be one record per line, and we probably should surround the object with curly braces.

In mTransformRecords, you print "Incoming SPI" and "Outgoing SPI" instead of the variable names. Since you are printing the SPIs in hex, you should add "0x" as a prefix.

In mTransformRecords, it looks like you print the pointer value of TransformRecord#mSocket. Instead, you should print the FD and port.

In mTransformRecords, you have a "Direction-IN" and a "Direction-Out" section, but it's not clear how they are grouped / related. Nested braces should help with that.

In mTransformRecords, you are mixing () and {} on the encryption= section.

Do you have a section for mSpiRecords?

Patch-set: 16
Reviewer: Gerrit User 1070023 <1070023@85c56323-6fa9-3386-8a01-6480fb634889>