Update patch set 2
Patch Set 2:

(1 comment)

I don't know how safe it is to make all these methods synchronized. For example, take interfaceLinkStateChanged. getLinkAddress involves an IPC to netd, which goes through the UNIX socket between NetworkManagementService and netd.

This could be blocked for several hundred milliseconds or more if a long-running netd command (e.g., an interface quota change) is running and/or has other commands backed up behind it.

Similarly, maybeSetIpv6NdOffload ends up sending an IPC to the wifi HAL, which has previously been known to block for up to several seconds.

Making everything in this class synchronized means that the interface observer threads can hold up the ConnectivityService handler thread (probably not too worrying) but also that the operations on the ConnectivityService thread (e.g., starting the clat daemon, which again involves sending an RPC to netd) can hold up BaseNetworkObserver notifications.

Holding up these notifications will delay operations that depend on them, such as IpManager provisioning, and cause failures in code (such as, I think, USB tethering) that depends on those notifications to be observed in a timely fashion.

I'm not sure if it's a good idea to introduce that sort of contention issue at this point.

I'd be more comfortable if we could avoid synchronization and just turn these functions into lambdas that we can just post to a handler thread (maybe even the ConnectivityService handler thread, assuming that's safe; I don't know how critical it is for that thread not to be blocked, particularly in the case where we need to acquire the NetworkManagementService network policy locks).

Another option would be to postpone this change until we move this code to IpManager. We have not seen this particular crash before, so it might be possible that it's so rare that we can afford not to fix it immediately.

Patch-set: 2
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>