OMS: add security checks for user-installed overlays
Restrict which overlays the system approves for use: for security
reasons, don't allow user-installed overlays unless both target and
overlay package are signed with the same certificate (which means they
share the same author, which means the author could have made the
modifications to the target package directly).

Pre-installed overlays are always permitted, regardless of signature.

Test: bit -t FrameworksServicesTests:com.android.server.om.{OverlayManagerSettingsTests,OverlayManagerTests,RuntimeResourceOverlayTests}
Change-Id: Ib58a5bdb15cb84f823658841abad77e12ed1ddc6