Connectivity: start PAC global proxy after reboot.
Previously, we read the global proxy settings from Settings.Global
database after reboot. This works for manual proxies, but not for
proxies based on PAC rules, as PacManager is never invoked when
ConnectivityService starts up. As a consequence, services required for
PAC-based proxies (PacService and ProxyService) are never started.

In this CL, we make sure PacManager is called during boot, so that PAC
rules will take effect. A broadcast is also sent to notify subscribers
that a new proxy rule is effective once its ready.

Bug: 35943997
Test: set a PAC proxy and observe it still works after reboot
Test: runtest frameworks-net
Change-Id: I69edb5fcd3e84663d338ff4b93c40fd5ee2c8c51