ConnectivityServiceTest: fix tearDown
tearDown now includes a final waitForIdle which ensures that
NetworkAgent disconnections are cleared from the various message queues
involved. This was causing testNetworkInfoOfTypeNone to fail some of the
time due to side effects from previously running tests in the same
instance of ConnectivityServiceTest.

Slso fixes some naming and typos in testNetworkInfoOfTypeNone

Bug: 62918393, 62918393
Test: runtest frameworks-net
Change-Id: I1c56557ab113d3ef57dbc06a6e882634d03c5b09