NET_CAPABILITY_NOT_METERED is mutable
This patch makes NET_CAPABILITY_NOT_METERED a "mutable" capability so
that it can be added or removed to NetworkCapabilities objects legally
by NetworkAgent implementations after they declare themselves connected.

This requires adding NET_CAPABILITY_NOT_METERED as an explicit exemption
so that apps can request it in NetworkRequests (checked by the unit test
ConnectivityServiceTest#testNetwokrFactoryRequests()).

This patch also improves the wtf() logging in updateCapabilities to
better distinguish between the cases of a changed specifiers, changed
transports, or changed capabilities.

Bug: 63326103
Test: runtest frameworks-net, runtest frameworks-wifi
Change-Id: I05c6e78891e1eac658f1cf883223af520a9a4f8f