Update patch set 1
Patch Set 1: Code-Review-2

The root issue is a driver bug and driver issues should go to the driver vendor. If the driver does not support EGL_KHR_partial_update it should not be returning that it supports that extension. HWUI's general policy is that it does not workaround driver bugs, as we aim to improve the state of the ecosystem which means forcing bugs to be fixed.

As you discovered there's the short-term workaround of debug.hwui.use_buffer_age=false that can be used until a fixed driver is acquired from the GPU vendor.

HWUI's usage of these extensions is correct and to spec, though.

Patch-set: 1
Reviewer: Gerrit User 1010103 <1010103@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2