Update patch set 1
Patch Set 1:

> Hi, thanks for the contribution.  Unfortunately we can't just
 > change this back as-is, since apps may have started to rely on the
 > behavior.  So it probably needs to have a targetSdkVersion check
 > for the next release, whatever that is.

I'm not sure what the best approach is to fix this.  Currently this caused issues when my app hit prod as our ramp up only had users with 7.0+ phones and then we had to ramp down as a result of crashes appearing once users with older phones hit the mix.  Implementing this fix without changing old versions can cause the same issues to rise until 7.0 -> this fix becomes dated and developers aren't testing on 7.0 phones.

If there is some way to at least document it so when you are developing you know of the issue that would fix a lot of the headaches.

Patch-set: 1
Reviewer: Gerrit User 1213978 <1213978@85c56323-6fa9-3386-8a01-6480fb634889>