NsdManager: allow callers to specify a custom Handler
This patch overloads registerService(), discoverService() and
resolveService() with versions that take an additional Handler argument.

By default listeners registered by the caller for these three apis are
run on a hidden background thread associated with NsdManager. If a
non-null Handler argument is provided, the listener is instead triggered
on that Handler.

Bug: 33298084
Bug: 62044295
Test: runtest frameworks-net
Change-Id: I6c97c088110604c5eda6035cdad82c0101e71b95