Update patch set 2
Patch Set 2:

> BTW, can a user bypass that quotas by making more than one IpSecService?

The user can bypass the quotas, but they can only bypass them on a separate IpSecService instance. The quota trackers are per-instance, but since a malicious app theoretically won't get access to NetD, *should* be stuck with an empty service and have no way to allocate objects that it tracks.

Would it be safer to convert IpSecService to a singleton pattern?

Patch-set: 2
Reviewer: Gerrit User 1060151 <1060151@85c56323-6fa9-3386-8a01-6480fb634889>