Update patch set 3
Patch Set 3:

> The user can bypass the quotas, but they can only bypass them on a separate IpSecService instance. The quota trackers are per-instance, but since a malicious app theoretically won't get access to NetD, *should* be stuck with an empty service and have no way to allocate objects that it tracks.
Would it be safer to convert IpSecService to a singleton pattern?

Not necessarily because if these is still a solution to create more than one instance via reflection on some other trick like class loader, then singleton does not help you.

Something easier to do, that we don in ConnectivityManager, is to put that state in a global static variable, with the obvious disadvantage that it makes testing harder. In any case, any validation needs to happen on the other side of binder too. Getting useful errors on the client side of Binder is also very useful anyway.

Patch-set: 3
Reviewer: Gerrit User 1107893 <1107893@85c56323-6fa9-3386-8a01-6480fb634889>