Clear insets from previous IME
When an IME is on screen and a new IME is chosen, the previous IME
insets are not cleared.  An issue can occur where the new IME does
not display a traditional IME UI, e.g. displays an AlertDialog.
This can result in the framework not updating the area of the screen
which is still considered to be an IME according to the previous insets.

This commit makes sure that the insets for the previous IME is always
cleared when a new IME is displayed.

Test: manual
Change-Id: I90f479050be3aca764eae8503697f31a37493ad6