Quick click twice to start Activity, it will open two Activities easily.
The original logic in Android N is:
when start a new Activity, it will set the previous Activity's visible false with setVisibility.
And then it should update the window's visible as false to inputflinger.

However, the isVisibleLw is not consider hiddenRequested in O.
so the window in inputflinger will be still visible after setVisibility with false.
And then the inputmotion will still can pass to Activity.

Fix:use isWinVisibleLw instead of isVisibleLw when passing the visible status to inputflinger.