Null pointer exception in FileRotator.java
Sometimes a null pointer exception is thrown when examining files
managed by the file rotator.

The logs from the test show that the Wifi state is changed a large
number of times. On every state change, a write operation is
initiated on the file system. This will eventually result in out
of memory and the call to mBasePath.list()method in FileRotator.java
will return null so the iteration will throw a NullPointerException.

Change-Id: Ib9c6ac16d3de5b3d4ea6c68d584c4bf04d60baf9
Signed-off-by: liangweikang@xiaomi.com <liangweikang@xiaomi.com>