Re-add safemode reboot confirmation dialog
When longpressing on power or restart, the device will
reboot into safemode without confirmation.  This can
result in users finding that they are 'stuck' in safemode
as it is not clear that tapping and long-pressing on the
buttons result in different outcomes.

This commit re-enables the confirmation dialog that was
removed when converting the global actions menu to a system ui
plugin.

Bug: 65037488
Test: Manual - longpress power/restart for confirmation
Change-Id: Iaf45e07c296ffebfa00a7ed552381a8ba9844925