Refactored to remove refs in releaseResources
Removed calls to remove references from managed arrays, and moved it to
releaseResources.

Considerations:
While binder death path and releaseManagedResources path both call
ManagedResource.release(), the release method itself has no way of
knowing which resource array to use. Thus, releaseResources is the
earliest point in which this can be done.

Also, the reference is removed as the last action of releaseResources,
to maintain the overall call order as much as possible.

Bug: 63409385
Change-Id: I90ccdeace925aa08f189aa47e9f20889297b834e
Test: New tests added to ensure cleanup when binderDied() is called