Update patch set 5
Patch Set 5:

I was made aware that the deprecation is to remove something that shouldn't have been public.  Its reasonable to say then that it is deprecated and that it does not function correctly, so shouldn't be used.  Whatever the case is, just make it clear in the comments.

Patch-set: 5
Reviewer: Gerrit User 1056478 <1056478@85c56323-6fa9-3386-8a01-6480fb634889>