Nat464Xlat: add STOPPING state
This patch adds a STOPPING state to Nat464Xlat in order to distinguish
the two following teardown scenarios:
 - network disconnects -> ConnectivityService calls Nat464Xlat#stop()
                       -> Nat464Xlat calls stopClatd
                       -> clatd stops and stacked interface is removed
                       -> Nat464Xlat#interfaceRemoved() is triggered
 - clatd crashes or exit -> Nat464Xlat#interfaceRemoved() is triggered
                         -> ConnectivityService is updated about the
                            stacked interface missing, and restarts
                            Nat464Xlat if needed.

Because in the second scenario stopClatd must be called for proper
cleanup, it forces the first scenario to have two possible calls to
stopClatd (one in stop() and one in interfaceRemoved()) which is
incorrect.

Furthermore, in the first scenario ConnectivityService should not be
updated about the stacked interface having been removed, because the
associated NetworkAgentInfo has already been cleaned up by
ConnectivityService.

Adding a STOPPING state that is entered in Nat464Xlat#stop() allows to
prevent both issues from happening in Nat464Xlat#interfaceRemoved().

Bug: 65225023
Bug: 64571917
Test: runtest frameworks-net
Change-Id: I27221a8a60fd9760b567ed322cc79228df877e56