Nat464Xlat: correct teardown, add unit tests
This patch disentangle stop() and handleInterfaceRemoved() so that
whichever is called first, Nat464Xlat terminates gracefully and reaches
a new STOPPED state. In STOPPED state, both stop() and
handleInterfaceRemoved() are no-op.

The common teardown procedure is put into enterStoppedState(). This
allows to clearly distinguish with stop() and interfaceRemoved() which
of the two following teardown scenarios happened:

 - network disconnects -> ConnectivityService calls Nat464Xlat#stop()
                       -> Nat464Xlat calls stopClatd
                       -> clatd stops and stacked interface is removed
                       -> Nat464Xlat#interfaceRemoved() is triggered and
                          ignored.

 - clatd crashes or exit -> Nat464Xlat#interfaceRemoved() is triggered
                         -> ConnectivityService is updated about the
                            stacked interface missing, and restarts
                            Nat464Xlat if needed.

Note that the first scenarios has two cases: stop() can be called before
the notification for the stacked interface going up (STARTED state), or
after (RUNNING state). Both are correctly handled byenterStoppedState().

Bug: 65225023
Bug: 64571917
Test: runtest frameworks-net
Change-Id: I27221a8a60fd9760b567ed322cc79228df877e56