Update patch set 3
Patch Set 3:

(1 comment)

What was your reasning for choosing to have the API call onResult() instead of onError() in the end?

If it's going to be onResult(null), it would be good to have the implementations inside ActivityManagerService mark it as @Nullable so we don't get NPE bugs in future. Right now they seem to be correctly defensive against it.

Patch-set: 3
Reviewer: Gerrit User 1045544 <1045544@85c56323-6fa9-3386-8a01-6480fb634889>