Update patch set 2
Patch Set 2:

I just found out about this patch.

Yes Paul is right, this is not enough synchronization. We got internal reports recently about this issue and are addressing it by moving all logic to ConnectivityService's handler.

This is done in patches:
https://android-review.googlesource.com/#/c/platform/frameworks/base/+/424680/ (merged)
https://android-review.googlesource.com/#/c/platform/frameworks/base/+/426280/ (merged)
https://android-review.googlesource.com/#/c/platform/frameworks/base/+/475397/ (soon to be merged)

We found several other races that we tried to all address in these patches.

Patch-set: 2
CC: Gerrit User 1107893 <1107893@85c56323-6fa9-3386-8a01-6480fb634889>