Update patch set 7
Patch Set 7: Code-Review+1

(8 comments)

Thank you for your persistence on this. This is a substantial improvement in the reliability of this code.

It looks like this CL is nearly there except for one issue (not unregistering the observer when we call stop() in STARTING state and an interfaceLinkStateChanged never happens) that should hopefully be straightforward to fix. See comments inline.

Patch-set: 7
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1