Added per process tracking of light state on InputDevices
Keep track of light state on input devices against a
Binder token.  This allows each app to function as if it
has total control of the lights, without disturbing other
applications' usage of them.  The current owner of the
present light state is also tracked.  When an application
attempts to read or change the state of the lights, the
lights are set to match that application's specific light
state.

Default light state of an input device is stored when the
device is added.  This allows the input device's lights
to be restored to their default state when the current
state's application exits.

Test: Start with the same test setup as Change-Id
      I3e694900c85e96900e6e03db2f13e809af6bca02.  If you kill
      the application the lights should return to their
      default state.  Run an app which uses the lights and
      then switch to a second app which also changes the
      lights.  Switch back to the first application and the
      first state request should set the lights back to that
      application's state.
Bug: 70356237
Change-Id: I524f59b82966965437348ab678eada0799241300