Support ENVELOPE (EVENT DOWNLOAD - User activity) command
The terminal shall send the ENVELOPE (EVENT DOWNLOAD - User activity)
command to the UICC when some user activity like a key-press or touch is
detected if the user activity event is a part of the current event list.
This change allows the SIM Toolkit application to support that event.
Please refer to the specification of SET UP EVENT LIST command described
in ETSI TS 102.223.

Test: Confirmed the expected seq 1.1 of ETSI TS 102.384 TC 27.22.4.16

Change-Id: I7d3faa09cd4625db53bcfe21f90570389a4ea036