Update patch set 1
Patch Set 1: Code-Review-2

This is WAI and should be fixed in apps that implement dragging in a ways that is affected by touch mode (ex. Launcher, which has already addressed the issue internally). Please file a bug on the AOSP Issue Tracker if you disagree, and we can discuss there.

Patch-set: 1
Reviewer: Gerrit User 1044446 <1044446@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2