Update patch set 1
Patch Set 1:

"-2" is because we're trying to get it fixed in a OC critical fix; and if that happens this CL will be dropped as part of a regular AOSP drop (after more testing).

This CL is published as "FYI" at this point.

Patch-set: 1
Reviewer: Gerrit User 1003801 <1003801@85c56323-6fa9-3386-8a01-6480fb634889>