Update patch set 1
Patch Set 1:

This should be a benign change, but let's discuss on b/33122276 to make sure this is really what you want. My understanding is this is "max", and if the device is low on memory, the framework will start killing apps on background users anyway, which I guess is why it's not configurable.

Patch-set: 1
Reviewer: Gerrit User 1003801 <1003801@85c56323-6fa9-3386-8a01-6480fb634889>