Update patch set 3
Patch Set 3:

> The mappings here *must* match the Android published standard, not
 > the Linux one, and this means that the left analog stick produces X
 > / Y, the right analog stick produces Z / RZ, and the triggers
 > produce LTRIGGER / RTRIGGER. Any other mapping won't be accepted.
 > 
 > 
 > > Please note that because the same mapping files are also used for
 > the motion sensors subdevice
 > 
 > Motion sensors are not an input device and we're unlikely to accept
 > a patch that makes them one; these should go through the dynamic
 > sensor APIs with some way to tie them to an input device id.

 > The mappings here *must* match the Android published standard, not
 > the Linux one, and this means that the left analog stick produces X
 > / Y, the right analog stick produces Z / RZ, and the triggers
 > produce LTRIGGER / RTRIGGER. Any other mapping won't be accepted.
 > 
 > 
 > > Please note that because the same mapping files are also used for
 > the motion sensors subdevice
 > 
 > Motion sensors are not an input device and we're unlikely to accept
 > a patch that makes them one; these should go through the dynamic
 > sensor APIs with some way to tie them to an input device id.

Motion sensors are a difficult topic. In case of the dualshock 3 / 4 they are part of the HID device and the data is intermingled in the input reports used for buttons and sticks.

We had long discussion on the linux-input kernel mailing list on how to proceed. Ultimately they really wanted to threat devices with motion sensors like the dualshock 3 / 4 as regular input devices. We were asked to report these through a composite device with INPUT_PROP_ACCELEROMETER set. This is now the recommended approach for motion sensors part of input devices. The code has supported since Linux 4.12 kernel.

https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux-stable.git/tree/Documentation/input/event-codes.rst

https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux-stable.git/tree/drivers/hid/hid-sony.c

Patch-set: 3
Reviewer: Gerrit User 1154383 <1154383@85c56323-6fa9-3386-8a01-6480fb634889>