Update patch set 3
Patch Set 3:

> > Ultimately they really wanted to threat devices with motion
 > sensors like the dualshock 3 / 4 as regular input devices.
 > 
 > That's really unfortunate since it means that standalone sensors go
 > through IIO and non-standalone sensors go through evdev. It would
 > be nice if they all came through IIO and they just also exposed a
 > handle of which input device they were associated with (if any).
 > 
 > Are there strong reasons why input device sensors are fundamentally
 > different from normal sensors? If not, then the kernel decision
 > just means that the sensor HAL will need to monitor and interpret
 > input devices in addition to its current source of sensor input.

We did a thorough investigation and it feels like there is no reason why the sensor framework wouldn't work. Previously it lacked important features, but at least of 7.0 there is support for dynamic devices.

If we went the sensor way, my understanding would be to:
- Add 'event dev based' sensors to libhardware/modules/sensors/dynamic_sensor. This can be done in a similar way how existing Dummy and HidRaw are handled. Does this sound reasonable?
- Filter out input devices with INPUT_PROP_ACCELEROMETER from inputflinger.

We should be able to drop existing linked reviews, which add some new types like a motion sensor input source to the input framework for handling input motion sensors.

The only big unknown is how to tie a sensor to InputDevice on the java side. My feeling is to expose a 'Sensor' (or list of sensors) on it similar how InputDevice exposes a Vibrator. We just don't know how to bubble the Sensor up to the InputDevice class. Any thoughts?

Patch-set: 3
Reviewer: Gerrit User 1154383 <1154383@85c56323-6fa9-3386-8a01-6480fb634889>