Update patch set 3
Patch Set 3:

> Add 'event dev based' sensors to libhardware/modules/sensors/dynamic_sensor. This can be done in a similar way how existing Dummy and HidRaw are handled. Does this sound reasonable?

pengxu@, can you help answer this? I'm not sure if adding stuff to the old HALs is still the right thing to do in a Treble-ized world.

> Filter out input devices with INPUT_PROP_ACCELEROMETER from inputflinger.

We don't actually want to filter out all input devices that set this, we just want to make sure we don't map their X / Y / Z axes to anything. They might have other valid forms of input still, which we don't want to ignore just because they happen to include an accelerometer.

> The only big unknown is how to tie a sensor to InputDevice on the java side. My feeling is to expose a 'Sensor' (or list of sensors) on it similar how InputDevice exposes a Vibrator. We just don't know how to bubble the Sensor up to the InputDevice class. Any thoughts?

In the past my thinking was that we'd have a getSensorManager() method that would know to only return sensors that are associated with this input device. Peng, any opinions on how we should expose this?

Patch-set: 3
Reviewer: Gerrit User 1006497 <1006497@85c56323-6fa9-3386-8a01-6480fb634889>