Check emergency or not on both subscriptions.
Check emergency or not on both subscriptions when modem is in
single standby. In current implementation we check for emergency
only on voice preferred subscription. However if modem is in
single stand by( sim not inserted in one/both of the slots or
pin/puk locked for one/both slots ), then we have to check both
ecc list of both subscriptions to know whether its emergency or
not and dial should be placed on active subscription.

Bug: 66897339
Change-Id: I97770468849e20f8ea9611ab19da50e40efb911b