Update patch set 3
Patch Set 3: Code-Review-1

If I'm reading this change correctly, it's incorrect because it only checks when tethering is enabled. That means that a user can enable tethering using a SIM card from a carrier that does not impose this restriction, then switch SIM cards and evade this restriction.

Additionally: from an API perspective it's not clear that a carrier should be able to restrict the user of the device from tethering between interfaces that are not controlled by the carrier (for example, turn on wifi hotspot when an USB ethernet adapter is plugged in, or enable USB tethering when the device is on wifi). I think we need a more nuanced approach here.

Patch-set: 3
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1