Update patch set 3
Patch Set 3: Code-Review+1

Looks okay to me. As far as this statistic goes, the fix is expected to be in for v24. This is v23, so that means the fix here is in early. Since the v23 statistic is not trusted anyway, it shouldn't matter that it is now suddenly fixed.

Patch-set: 3
Reviewer: Gerrit User 1165283 <1165283@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1