Track moving of OkHttp facade API to new package.
Another CL in this topic moves the classes that make up this
facade from libcore.net.http to com.squareup.okhttp.facadeapi; in
jarjar'ed build targets, this becomes com.android.okhttp.facadeapi.

This facade constitutes the API via which Android framework may
access OkHttp. It's moving because libcore.net.http is already
part of libcore, and the overlap of packages is problematic for
builds with EXPERIMENTAL_USE_OPENJDK9 set to true.

Bug: 68220880
Test: Treehugger
Change-Id: Ia79966563cc0b5ab0923d54c21e54b6192d8c990